+++ Rag_5040/MNK.lua
@@ -29,6 +29,9 @@ local kampfer_ring = {
 local kampfer_earring = {
     Ear2 = 'Kampfer Earring',
 }
+local shadow_mantle = {
+    Back = 'Shadow Mantle',
+}
 
 local sets = {
     Idle = {
@@ -295,9 +298,9 @@ sets.temple_cyclas = temple_cyclas
 sets.temple_crown = temple_crown
 sets.melee_gaiters = melee_gaiters
 sets.melee_gloves = melee_gloves
-
 sets.kampfer_ring = kampfer_ring
 sets.kampfer_earring = kampfer_earring
+sets.shadow_mantle = shadow_mantle
 profile.Sets = gcmelee.AppendSets(sets)
 
 profile.HandleAbility = function()
@@ -313,6 +316,11 @@ profile.HandleAbility = function()
         gFunc.EquipSet(sets.Chakra)
         gFunc.EquipSet('temple_cyclas')
         gFunc.EquipSet('melee_gloves')
+
+        local environment = gData.GetEnvironment()
+        if (environment.DayElement == 'Dark') then
+            gFunc.EquipSet('shadow_mantle')
+        end
     elseif (action.Name == 'Dodge') then
         gFunc.EquipSet('temple_gaiters')
     elseif (action.Name == 'Boost') then