+++ Rag_5040/SMN.lua
@@ -37,6 +37,25 @@ local bahamuts_staff = {
     -- Main = 'Bahamut\'s Staff',
 }
 
+-- Disabled on horizon_safe_mode
+local conjurers_earring_hp_threshold = 360 -- HP at which Conjurer's Earring set is equipped
+local conjurers_earring = { -- 
+    Main = 'Terra\'s Staff',
+    Ammo = 'Hedgehog Bomb',
+    -- Head = 'Example',
+    Neck = 'Pch. Collar',
+    Ear1 = 'Cassie Earring',
+    Ear2 = 'Conjurer\'s Earring',
+    -- Body = 'Example',
+    Hands = 'Garden Bangles',
+    Ring1 = 'Bomb Queen Ring',
+    Ring2 = 'Sattva Ring',
+    Back = 'Gigant Mantle',
+    Waist = 'Ocean Sash',
+    -- Legs = 'Example',
+    Feet = 'Marine M Boots',
+}
+
 local sets = {
     Idle = {
         Main = 'Terra\'s Staff',
@@ -78,6 +97,9 @@ local sets = {
 
     Perpetuation = { -- There is no point in using this set over an Idle set except for equipping Penance Robe
         -- Body = 'Penance Robe',
+    },
+    Perpetuation_Spirits = { -- For maximizing summoning skill with spirits out if you wish
+
     },
 
     DT = {
@@ -411,6 +433,7 @@ sets.summoners_doublet = summoners_doublet
 sets.summoners_horn = summoners_horn
 sets.conjurers_ring = conjurers_ring
 sets.bahamuts_staff = bahamuts_staff
+sets.conjurers_earring = conjurers_earring
 profile.Sets = gcmage.AppendSets(sets)
 
 -- Includes Chaotic Strike and Shock Strike in SmnSkill to maximize stun chance
@@ -465,7 +488,9 @@ profile.HandleCommand = function(args)
 end
 
 profile.HandleDefault = function()
+    local player = gData.GetPlayer()
     local petAction = gData.GetPetAction()
+
     if (petAction ~= nil) then
         gFunc.EquipSet('BP')
 
@@ -491,7 +516,7 @@ profile.HandleDefault = function()
     else
         if (not gcinclude.horizon_safe_mode) then
             local player = gData.GetPlayer()
-            if (conjurersRingForced and player.HP >= conjurersRingMaxHP) then
+            if (gcdisplay.GetToggle('C-Ring') and player.HP >= conjurersRingMaxHP and gData.GetPet()) then
                 local time = os.clock()
                 if (time > nextConjurersRingCheck) then
                     nextConjurersRingCheck = time + 3 -- only recheck again after 3 seconds to prevent spam
@@ -502,6 +527,7 @@ profile.HandleDefault = function()
         end
 
         gcmage.DoDefault(sets, nil, nil, nil, nil, nil)
+        gcmage.DoDefaultOverride()
     end
     gFunc.EquipSet(gcinclude.BuildLockableSet(gData.GetEquipment()))
 end