+++ Rag_5040/WHM.lua
@@ -37,6 +37,25 @@ local asklepios = { -- Used for Cures with Mjollnir when /NIN
     -- Sub = 'Asklepios',
 }
 
+-- Disabled on horizon_safe_mode
+local medicine_earring_hp_threshold = 360 -- HP at which Medicine Earring set is equipped
+local medicine_earring = { -- 
+    Main = 'Terra\'s Staff',
+    Ammo = 'Hedgehog Bomb',
+    -- Head = 'Example',
+    Neck = 'Pch. Collar',
+    Ear1 = 'Cassie Earring',
+    Ear2 = 'Medicine Earring',
+    -- Body = 'Example',
+    Hands = 'Garden Bangles',
+    Ring1 = 'Bomb Queen Ring',
+    Ring2 = 'Sattva Ring',
+    Back = 'Gigant Mantle',
+    Waist = 'Ocean Sash',
+    -- Legs = 'Example',
+    Feet = 'Marine M Boots',
+}
+
 local sets = {
     Idle = {},
     IdleALT = {},
@@ -143,6 +162,7 @@ sets.ruckes_rung = ruckes_rung
 sets.medicine_ring = medicine_ring
 sets.mjollnir = mjollnir
 sets.asklepios = asklepios
+sets.medicine_earring = medicine_earring
 profile.Sets = gcmage.AppendSets(sets)
 
 profile.HandleAbility = function()
@@ -199,6 +219,14 @@ end
 
 profile.HandleDefault = function()
     gcmage.DoDefault(sets, ninSJMaxMP, nil, blmSJMaxMP, rdmSJMaxMP, drkSJMaxMP)
+
+    local player = gData.GetPlayer()
+    if (not gcinclude.horizon_safe_mode) then
+        if (player.HP <= medicine_earring_hp_threshold) then
+            gFunc.EquipSet('medicine_earring')
+        end
+    end
+
     gcmage.DoDefaultOverride()
     gFunc.EquipSet(gcinclude.BuildLockableSet(gData.GetEquipment()))
 end
@@ -234,6 +262,8 @@ profile.HandleMidcast = function()
             end
         end
     end
+
+    gcmage.EquipWeaponLoadout()
 end
 
 return profile